+++ postpic/io/__init__.py
@@ -21,9 +21,10 @@
 The postpic.io module provides free functions for importing and exporting data.
 '''
 
-from .csv import _export_field_csv
+from .csv import _export_field_csv, _import_field_csv
 from .npy import _export_field_npy, _import_field_npy
 from .vtk import export_scalar_vtk, export_vector_vtk, export_scalars_vtk
+from .image import _import_field_image
 
 __all__ = ['export_field', 'load_field',
            'export_scalar_vtk', 'export_scalars_vtk', 'export_vector_vtk']
@@ -31,13 +32,29 @@ __all__ = ['export_field', 'load_field',
 
 def load_field(filename):
     '''
-    construct a new field object from file. currently, the following file
-    formats are supported:
-    *.npz
+    Load a field object previously stored using the `saveto` method. These are .npz files
+    with a specific metadata layout.
     '''
-    if not filename.endswith('npz'):
+    if not filename.lower().endswith('npz'):
         raise ValueError('File format of filename {0} not recognized.'.format(filename))
     return _import_field_npy(filename)
+
+
+def import_field(filename, **kwargs):
+    '''
+    Construct a new field object from foreign data. Currently, the following file
+    formats are supported explicitly:
+    *.csv
+    *.png
+
+    All other files will be opened using Pillow.
+    '''
+    if filename.lower().endswith('csv'):
+        return _import_field_csv(filename, **kwargs)
+    else:
+        # assume anything else, this will open png files with pypng and all other files
+        # with pillow
+        return _import_field_image(filename, **kwargs)
 
 
 def export_field(filename, field, **kwargs):
@@ -51,11 +68,11 @@ def export_field(filename, field, **kwargs):
     .vtk:
         vtk export to paraview
     '''
-    if filename.endswith('npz'):
+    if filename.lower().endswith('npz'):
         _export_field_npy(filename, field, **kwargs)
-    elif filename.endswith('csv'):
+    elif filename.lower().endswith('csv'):
         _export_field_csv(filename, field, **kwargs)
-    elif filename.endswith('vtk'):
+    elif filename.lower().endswith('vtk'):
         export_scalar_vtk(filename, field, **kwargs)
     else:
         raise ValueError('File format of filename {0} not recognized.'.format(filename))